backend/requirements.txt

django
djangorestframework
django-cors-headers
supabase
pdf2docx
pypdf
Pillow
python-docx
